# src/za.co.grindrodbank.a3s/wwwroot/openapi-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "description" : "API Definition for the A3S. This service allows authentication, authorisation and accounting.",
    "title" : "A3S",
    "version" : "1.0.5"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/twoFactorAuth/users/{userId}" : {
      "delete" : {
        "description" : "Removes two-factor authentication for a user.",
        "operationId" : "removeTwoFactorAuthentication",
        "parameters" : [ {
          "description" : "The UUID of the user.",
          "explode" : false,
          "in" : "path",
          "name" : "userId",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UserId"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content."
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Bad Request."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Not authenticated."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Forbidden - You are not authorized to remove two-factor authentication."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "User not found."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "An unexpected error occurred."
          }
        },
        "summary" : "Removes two-factor authentication for a user, as well as any registered devices and methods.",
        "tags" : [ "TwoFactorAuth" ]
      }
    },
    "/twoFactorAuth/validate" : {
      "post" : {
        "description" : "Validates a two-factor authenticator OTP.",
        "operationId" : "validateTwoFactorAuthenticationOTP",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TwoFactorAuthOTP"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ValidationResultResponse"
                }
              }
            },
            "description" : "Successful. OTP validation ran successfully. Check Success parameter and results for more information."
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Bad Request."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Not authenticated."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Forbidden - You are not authorized to validate two-factor authentication OTP's."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "User not found."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "An unexpected error occurred."
          }
        },
        "summary" : "Validates a two-factor authenticator OTP.",
        "tags" : [ "TwoFactorAuth" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "UserId" : {
        "format" : "uuid",
        "type" : "string"
      },
...
      "TwoFactorAuthOTP" : {
        "description" : "A OTP for a specific user.\n",
        "example" : {
          "OTP" : "OTP",
          "userId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
        },
        "properties" : {
          "userId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "OTP" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ErrorResponse" : {
        "description" : "Error response object to give more info with HTTP status code",
        "properties" : {
          "message" : {
            "type" : "string"
          }
        },
        "required" : [ "message" ],
        "type" : "object"
      },
...
      "ValidationResultResponse" : {
        "description" : "A validation result.\n",
        "example" : {
          "success" : true,
          "messages" : [ "messages", "messages" ]
        },
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "messages" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      }
    }
  }
}